IDPT-802 Configure FittingCompressedSizeCollectionViewController
