+++ Example/ReactiveDataDisplayManager/Collection/Views/Cells/FittingCompressedSizeCollectionViewCell/DifferentSizeCollectionViewCellModel.swift
@@ -11,6 +11,5 @@ struct DifferentSizeCollectionViewCellModel {
 
     let title: String
     let backgroundColor: UIColor
-    let height: CGFloat
 
 }
