+++ Example/ReactiveDataDisplayManager/Collection/Views/Cells/FittingCompressedSizeCollectionViewCell/FittingCompressedSizeCollectionViewCell.xib
@@ -10,7 +10,7 @@
     <objects>
         <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner"/>
         <placeholder placeholderIdentifier="IBFirstResponder" id="-2" customClass="UIResponder"/>
-        <collectionViewCell opaque="NO" clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="center" id="zI5-wR-32G" customClass="DifferentSizeCollectionViewCell" customModule="ReactiveDataDisplayManagerExample_iOS" customModuleProvider="target">
+        <collectionViewCell opaque="NO" clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="center" id="zI5-wR-32G" customClass="FittingCompressedSizeCollectionViewCell" customModule="ReactiveDataDisplayManagerExample_iOS" customModuleProvider="target">
             <rect key="frame" x="0.0" y="0.0" width="183" height="64"/>
             <autoresizingMask key="autoresizingMask" flexibleMaxX="YES" flexibleMaxY="YES"/>
             <collectionViewCellContentView key="contentView" opaque="NO" clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="center" insetsLayoutMarginsFromSafeArea="NO" id="3jZ-Ve-WUq">
@@ -18,17 +18,17 @@
                 <autoresizingMask key="autoresizingMask" flexibleMaxX="YES" flexibleMaxY="YES"/>
                 <subviews>
                     <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="Label" textAlignment="center" lineBreakMode="tailTruncation" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="cfe-Lx-79M">
-                        <rect key="frame" x="0.0" y="0.0" width="183" height="64"/>
+                        <rect key="frame" x="5" y="5" width="173" height="54"/>
                         <fontDescription key="fontDescription" type="system" pointSize="17"/>
                         <nil key="textColor"/>
                         <nil key="highlightedColor"/>
                     </label>
                 </subviews>
                 <constraints>
-                    <constraint firstItem="cfe-Lx-79M" firstAttribute="leading" secondItem="3jZ-Ve-WUq" secondAttribute="leading" id="1rm-uq-VvM"/>
-                    <constraint firstAttribute="bottom" secondItem="cfe-Lx-79M" secondAttribute="bottom" id="b6e-cX-uKI"/>
-                    <constraint firstItem="cfe-Lx-79M" firstAttribute="top" secondItem="3jZ-Ve-WUq" secondAttribute="top" id="fMg-GI-4SP"/>
-                    <constraint firstAttribute="trailing" secondItem="cfe-Lx-79M" secondAttribute="trailing" id="skm-oc-wM0"/>
+                    <constraint firstItem="cfe-Lx-79M" firstAttribute="leading" secondItem="3jZ-Ve-WUq" secondAttribute="leading" constant="5" id="1rm-uq-VvM"/>
+                    <constraint firstAttribute="bottom" secondItem="cfe-Lx-79M" secondAttribute="bottom" constant="5" id="b6e-cX-uKI"/>
+                    <constraint firstItem="cfe-Lx-79M" firstAttribute="top" secondItem="3jZ-Ve-WUq" secondAttribute="top" constant="5" id="fMg-GI-4SP"/>
+                    <constraint firstAttribute="trailing" secondItem="cfe-Lx-79M" secondAttribute="trailing" constant="5" id="skm-oc-wM0"/>
                 </constraints>
             </collectionViewCellContentView>
             <size key="customSize" width="183" height="64"/>
